Inline BasePublishPlugin
